backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
alpaca-py>=0.20.0
anthropic>=0.40.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
